Added I2C driver for Audio Codec
This is the updated driver the WM8731 codec. The writeup is out of data, but I left it in because it details some of the debugging processes that I used. The coded does now work as intended loading up a sof audio file into an IP ROM. The codec is then configured over I2C and plays the audio out of the ROM.